[RUST] Split series tests (#749)
* It is currently difficult to understand test coverage across types and
ops because everything is in one file
* Splits `test_series.py` into individual files for each family of
operations for easier contribution of tests
