+++ tests/series/test_cast.py
@@ -1,0 +1,42 @@
+from __future__ import annotations
+
+import itertools
+
+import pyarrow as pa
+import pytest
+
+from daft.datatype import DataType
+from daft.series import Series
+from tests.series import ARROW_FLOAT_TYPES, ARROW_INT_TYPES, ARROW_STRING_TYPES
+
+
+@pytest.mark.parametrize("source_dtype, dest_dtype", itertools.product(ARROW_INT_TYPES + ARROW_FLOAT_TYPES, repeat=2))
+def test_series_casting_numeric(source_dtype, dest_dtype) -> None:
+    data = pa.array([1, 2, 3, None, 5, None])
+
+    s = Series.from_arrow(data.cast(source_dtype))
+
+    assert s.datatype() == DataType.from_arrow_type(source_dtype)
+    target_dtype = DataType.from_arrow_type(dest_dtype)
+    t = s.cast(target_dtype)
+    assert t.datatype() == target_dtype
+    assert t.to_pylist() == [1, 2, 3, None, 5, None]
+
+
+@pytest.mark.parametrize(
+    "source_dtype, dest_dtype", itertools.product(ARROW_INT_TYPES + ARROW_FLOAT_TYPES, ARROW_STRING_TYPES)
+)
+def test_series_casting_to_string(source_dtype, dest_dtype) -> None:
+    data = pa.array([1, 2, 3, None, 5, None])
+
+    s = Series.from_arrow(data.cast(source_dtype))
+
+    assert s.datatype() == DataType.from_arrow_type(source_dtype)
+    target_dtype = DataType.from_arrow_type(dest_dtype)
+    t = s.cast(target_dtype)
+    assert t.datatype() == target_dtype
+
+    if pa.types.is_floating(source_dtype):
+        assert t.to_pylist() == ["1.0", "2.0", "3.0", None, "5.0", None]
+    else:
+        assert t.to_pylist() == ["1", "2", "3", None, "5", None]
