+++ tests/series/test_take.py
@@ -1,0 +1,40 @@
+from __future__ import annotations
+
+import pyarrow as pa
+import pytest
+
+from daft.datatype import DataType
+from daft.series import Series
+from tests.series import ARROW_FLOAT_TYPES, ARROW_INT_TYPES, ARROW_STRING_TYPES
+
+
+@pytest.mark.parametrize("dtype", ARROW_INT_TYPES + ARROW_FLOAT_TYPES + ARROW_STRING_TYPES)
+def test_series_take_numeric(dtype) -> None:
+    data = pa.array([1, 2, 3, None, 5, None])
+
+    s = Series.from_arrow(data.cast(dtype))
+    pyidx = [2, 0, None, 5]
+    idx = Series.from_pylist(pyidx)
+
+    result = s.take(idx)
+    assert result.datatype() == s.datatype()
+    assert len(result) == 4
+
+    original_data = s.to_pylist()
+    expected = [original_data[i] if i is not None else None for i in pyidx]
+    assert result.to_pylist() == expected
+
+
+def test_series_date_take() -> None:
+    from datetime import date
+
+    def date_maker(d):
+        if d is None:
+            return None
+        return date(2023, 1, d)
+
+    days = list(map(date_maker, [5, 4, 1, None, 2, None]))
+    s = Series.from_pylist(days)
+    taken = s.take(Series.from_pylist([5, 4, 3, 2, 1, 0]))
+    assert taken.datatype() == DataType.date()
+    assert taken.to_pylist() == days[::-1]
